adding two new migrations finish the rationg in the article api controller finish some other methods need to implement the comments later on

+++ Code/NewsWebsiteBackEnd/DTO/Article/GeneralArticleDetailsViewModel.cs
@@ -9,6 +9,6 @@
         public string Title { get; set; }
         public string CoverImagePath { get; set; }
         public int TotalNumberOfViews { get; set; }
-        public decimal RatingAvg { get; set; }
+        public double RatingAvg { get; set; }
     }
 }
